Refactored tasks
