+++ posthog/apps.py
@@ -7,6 +7,7 @@ from django.conf import settings
 from posthoganalytics.client import Client
 
 from posthog.settings import SELF_CAPTURE, SKIP_ASYNC_MIGRATIONS_SETUP
+from posthog.tasks.tasks import sync_all_organization_available_features
 from posthog.utils import (
     get_git_branch,
     get_git_commit,
@@ -35,8 +36,6 @@ class PostHogConfig(AppConfig):
             # log development server launch to posthog
             if os.getenv("RUN_MAIN") == "true":
                 # Sync all organization.available_features once on launch, in case plans changed
-                from posthog.celery import sync_all_organization_available_features
-
                 sync_all_organization_available_features()
 
                 # NOTE: This has to be created as a separate client so that the "capture" call doesn't lock in the properties
