+++ posthog/models/insight_caching_state.py
@@ -8,6 +8,7 @@ from posthog.models.sharing_configuration import SharingConfiguration
 from posthog.models.signals import mutable_receiver
 from posthog.models.team import Team
 from posthog.models.utils import UniqueConstraintByExpression, UUIDModel
+from posthog.tasks.tasks import sync_insight_caching_state
 
 
 class InsightCachingState(UUIDModel):
@@ -48,8 +49,6 @@ class InsightCachingState(UUIDModel):
 
 @mutable_receiver(post_save, sender=SharingConfiguration)
 def sync_sharing_configuration(sender, instance: SharingConfiguration, **kwargs):
-    from posthog.celery import sync_insight_caching_state
-
     if instance.insight is not None and not instance.insight.deleted:
         sync_insight_caching_state.delay(instance.team_id, insight_id=instance.insight_id)
     elif instance.dashboard is not None and not instance.dashboard.deleted:
@@ -59,22 +58,16 @@ def sync_sharing_configuration(sender, instance: SharingConfiguration, **kwargs)
 
 @mutable_receiver(post_save, sender=Insight)
 def sync_insight(sender, instance: Insight, **kwargs):
-    from posthog.celery import sync_insight_caching_state
-
     sync_insight_caching_state.delay(instance.team_id, insight_id=instance.pk)
 
 
 @mutable_receiver(post_save, sender=DashboardTile)
 def sync_dashboard_tile(sender, instance: DashboardTile, **kwargs):
-    from posthog.celery import sync_insight_caching_state
-
     sync_insight_caching_state.delay(instance.dashboard.team_id, dashboard_tile_id=instance.pk)
 
 
 @mutable_receiver(post_save, sender=Dashboard)
 def sync_dashboard_updated(sender, instance: Dashboard, **kwargs):
-    from posthog.celery import sync_insight_caching_state
-
     update_fields = kwargs.get("update_fields")
     if update_fields in [
         frozenset({"filters_hash"}),
