+++ posthog/tasks/__init__.py
@@ -10,6 +10,7 @@ from . import (
     process_scheduled_changes,
     split_person,
     sync_all_organization_available_features,
+    tasks,
     usage_report,
     user_identify,
     warehouse,
@@ -25,6 +26,7 @@ __all__ = [
     "process_scheduled_changes",
     "split_person",
     "sync_all_organization_available_features",
+    "tasks",
     "user_identify",
     "usage_report",
     "warehouse",
