+++ posthog/tasks/split_person.py
@@ -1,10 +1,11 @@
 from typing import Optional
 
-from posthog.celery import app
+from celery import shared_task
+
 from posthog.models import Person
 
 
-@app.task(ignore_result=True, max_retries=1)
+@shared_task(ignore_result=True, max_retries=1)
 def split_person(person_id: int, main_distinct_id: Optional[str], max_splits: Optional[int]) -> None:
     """
     Split all distinct ids into separate persons
