+++ posthog/tasks/usage_report.py
@@ -17,6 +17,7 @@ from typing import (
 
 import requests
 import structlog
+from celery import shared_task
 from dateutil import parser
 from django.conf import settings
 from django.db import connection
@@ -27,7 +28,6 @@ from retry import retry
 from sentry_sdk import capture_exception
 
 from posthog import version_requirement
-from posthog.celery import app
 from posthog.clickhouse.client.connection import Workload
 from posthog.client import sync_execute
 from posthog.cloud_utils import get_cached_instance_license, is_cloud
@@ -274,7 +274,7 @@ def get_org_owner_or_first_user(organization_id: str) -> Optional[User]:
     return user
 
 
-@app.task(ignore_result=True, autoretry_for=(Exception,), max_retries=3)
+@shared_task(ignore_result=True, autoretry_for=(Exception,), max_retries=3)
 def send_report_to_billing_service(org_id: str, report: Dict[str, Any]) -> None:
     if not settings.EE_AVAILABLE:
         return
@@ -621,7 +621,7 @@ def get_teams_with_rows_synced_in_period(begin: datetime, end: datetime) -> List
     return results
 
 
-@app.task(ignore_result=True, max_retries=0)
+@shared_task(ignore_result=True, max_retries=0)
 def capture_report(
     capture_event_name: str,
     org_id: str,
@@ -952,7 +952,7 @@ def _get_full_org_usage_report_as_dict(full_report: FullUsageReport) -> Dict[str
     return dataclasses.asdict(full_report)
 
 
-@app.task(ignore_result=True, max_retries=3, autoretry_for=(Exception,))
+@shared_task(ignore_result=True, max_retries=3, autoretry_for=(Exception,))
 def send_all_org_usage_reports(
     dry_run: bool = False,
     at: Optional[str] = None,
