+++ posthog/tasks/user_identify.py
@@ -1,10 +1,10 @@
 import posthoganalytics
+from celery import shared_task
 
-from posthog.celery import app
 from posthog.models import User
 
 
-@app.task(ignore_result=True)
+@shared_task(ignore_result=True)
 def identify_task(user_id: int) -> None:
     user = User.objects.get(id=user_id)
     posthoganalytics.capture(
