+++ posthog/tasks/verify_persons_data_in_sync.py
@@ -4,10 +4,10 @@ from datetime import timedelta
 from typing import Any, Dict, List
 
 import structlog
+from celery import shared_task
 from django.db.models.query import Prefetch
 from django.utils.timezone import now
 
-from posthog.celery import app
 from posthog.client import sync_execute
 from posthog.models.person import Person
 
@@ -41,7 +41,7 @@ HAVING argMax(is_deleted, version) = 0 AND person_id IN (%(person_ids)s)
 """
 
 
-@app.task(max_retries=1, ignore_result=True)
+@shared_task(max_retries=1, ignore_result=True)
 def verify_persons_data_in_sync(
     period_start: timedelta = PERIOD_START,
     period_end: timedelta = PERIOD_END,
